assignment handling
Improve how we deal with assignments. we now have three DataSet
structures:
 * one with full assignments. this can be synced to the server and has
 all info on the assignment
 * one with engineer assignments. this only has the info necessary to
 make the engineer timeline
 * one with project assignments. this only has the info necessary to
 make the project timeline

This is necessary because the timelines have requirements on the key
names for binding them to the plot. Each dataset uses the assignment id (aid)
as main index, to be used with get(). After modifying something, dont forget to
call update() on the DataSet.